Ring-buffer is now mutex-protected.
While an atomic compare-exchange could be used to update the start/end values, there is no assurance that the item will successfully be copied out of ring-buffer before it is overwritten. It also creates a race-condition for a consumer reading the buffer, while a producer is assigning a value. I think a mutex is the only way to do this.